uk-esos-app-api/src/test/java/uk/gov/esos/api/workflow/request/flow/rde/handler/RdeResponseSubmitActionHandlerTest.java
package uk.gov.esos.api.workflow.request.flow.rde.handler;

import static org.assertj.core.api.Assertions.assertThat;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import uk.gov.esos.api.workflow.request.WorkflowService;
import uk.gov.esos.api.workflow.request.core.domain.enumeration.RequestTaskActionType;
import uk.gov.esos.api.workflow.request.core.service.RequestTaskService;


@ExtendWith(MockitoExtension.class)
class RdeResponseSubmitActionHandlerTest {

    @InjectMocks
    private RdeResponseSubmitActionHandler handler;

    @Mock
    private RequestTaskService requestTaskService;

    @Mock
    private WorkflowService workflowService;

    @Test
    void getTypes() {
        assertThat(handler.getTypes()).containsExactly(RequestTaskActionType.RDE_RESPONSE_SUBMIT);
    }
}
